requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
streamlit>=1.25.0
requests>=2.25.0
pandas>=1.5.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0

# Database connectors
# sqlite3 is part of Python standard library
mysql-connector-python>=8.0.0
psycopg2-binary>=2.8.0
pymongo>=4.0.0

# Optional: for better JSON handling
# orjson>=3.8.0